refactor(playwright): extract embedded frontend route helper to shared module
`tests/index-status.spec.ts` の page.route handler / asset resolver /
contentType を `tests/_helpers/embedded-frontend.ts` に切り出し、spec
本体は WebSocket fixture と test 本文に集中させる。kanban.spec.ts は
別 owner なので触らず、index-status spec のみリファクタリング。

- 新 helper: `installEmbeddedRoutes(page)` + `APP_URL` を export
- spec から fixture-routing 系 helper (resolveAssetPath / contentTypeFor /
  ROOT_MODULES) を削除し、import で吸収
- spec が 742 → 661 lines に圧縮

26/26 (13 spec × 2 theme) pass を維持。production behavior 等価。
