+++ crates/gwt/playwright/tests/index-status.spec.ts
@@ -1,39 +1,13 @@
-/* SPEC-1939 Phase 12 / T-IDX-109 + T-IDX-110 — Project Index status badge.
+/* SPEC-1939 Phase 12 / T-IDX-102..T-IDX-110 — Project Index status badge.
  *
- * Reuses the SPEC-2017 Kanban fixture pattern (`tests/kanban.spec.ts`):
- * the fixture serves embedded frontend assets through Playwright routes
- * and stubs WebSocket with a deterministic backend that emits canned
+ * Reuses the SPEC-2017 Kanban fixture pattern: the embedded frontend is
+ * served via `installEmbeddedRoutes` (`_helpers/embedded-frontend.ts`) and
+ * the WebSocket is stubbed with a deterministic backend that emits canned
  * `workspace_state` + `project_index_status` events. No xvfb / wry / live
  * gwt process required, so the suite stays reliable in headless CI.
  */
 import { expect, test } from "@playwright/test";
-import path from "node:path";
-import { fileURLToPath } from "node:url";
-
-const APP_URL = "http://gwt-playwright.local/";
-const WEB_ROOT = path.resolve(
-  path.dirname(fileURLToPath(import.meta.url)),
-  "../../web",
-);
-
-const ROOT_MODULES = new Set([
-  "app.js",
-  "board-surface.js",
-  "branch-cleanup-modal.js",
-  "focus-trap.js",
-  "hotkey.js",
-  "index-settings-panel.js",
-  "index-status-controller.js",
-  "migration-modal.js",
-  "operator-shell.js",
-  "terminal-context-menu.js",
-  "terminal-viewport-reflow.js",
-  "theme-manager.js",
-  "theme-toggle.js",
-  "update-cta.js",
-  "window-docking.js",
-  "workspace-kanban-surface.js",
-]);
+import { APP_URL, installEmbeddedRoutes } from "./_helpers/embedded-frontend";
 
 test.describe("Project Index status badge", () => {
   test.use({ viewport: { width: 1440, height: 900 } });
@@ -541,61 +515,6 @@ test.describe("Project Index status badge", () => {
     await expect(toast).toContainText(/1 of 4 scope/);
   });
 });
-
-async function installEmbeddedRoutes(page) {
-  await page.route("http://gwt-playwright.local/**", async (route) => {
-    const url = new URL(route.request().url());
-    const assetPath = resolveAssetPath(url.pathname);
-    if (!assetPath) {
-      await route.fulfill({
-        status: 404,
-        contentType: "text/plain",
-        body: `No test asset for ${url.pathname}`,
-      });
-      return;
-    }
-    await route.fulfill({
-      path: assetPath,
-      contentType: contentTypeFor(assetPath),
-    });
-  });
-}
-
-function resolveAssetPath(pathname) {
-  if (pathname === "/" || pathname === "/index.html") {
-    return path.join(WEB_ROOT, "index.html");
-  }
-  if (pathname === "/assets/xterm/xterm.css") {
-    return path.join(WEB_ROOT, "vendor/xterm/xterm.css");
-  }
-  if (pathname === "/assets/xterm/xterm.mjs") {
-    return path.join(WEB_ROOT, "vendor/xterm/xterm.mjs");
-  }
-  if (pathname === "/assets/xterm/addon-fit.mjs") {
-    return path.join(WEB_ROOT, "vendor/xterm/addon-fit.mjs");
-  }
-  if (pathname.startsWith("/assets/fonts/")) {
-    return path.join(WEB_ROOT, "fonts", path.basename(pathname));
-  }
-  if (pathname.startsWith("/styles/")) {
-    return path.join(WEB_ROOT, "styles", path.basename(pathname));
-  }
-  const moduleName = pathname.slice(1);
-  if (ROOT_MODULES.has(moduleName)) {
-    return path.join(WEB_ROOT, moduleName);
-  }
-  return null;
-}
-
-function contentTypeFor(assetPath) {
-  if (assetPath.endsWith(".html")) return "text/html";
-  if (assetPath.endsWith(".css")) return "text/css";
-  if (assetPath.endsWith(".js") || assetPath.endsWith(".mjs")) {
-    return "text/javascript";
-  }
-  if (assetPath.endsWith(".woff2")) return "font/woff2";
-  return "application/octet-stream";
-}
 
 async function installIndexStatusBackend(page, indexStatus) {
   await page.addInitScript((indexStatusPayload) => {
